smarter way to handle cover art

+++ scrobble_shelf/scrobble_shelf.py
@@ -1,11 +1,10 @@
 #! /usr/bin/env python
 import pylast, os, json, argparse, requests, mimetypes, urllib
+from .cover_art_subs import COVER_ART_SUBS
 from shutil import copyfileobj
 
 API_KEY = os.environ['LAST_FM_API_KEY']
 API_SECRET = os.environ['LAST_FM_API_SECRET']
-
-COVER_ART_SUBS = {} 
 
 parser = argparse.ArgumentParser(description='Create interactive album collages from last.fm scrobble stats.')
 parser.add_argument('username')
@@ -33,18 +32,17 @@ class ScrobbleShelf():
                 if cover_art_url:
                     cover_art = requests.get(album.get_cover_image(), stream=True)
                     extension = mimetypes.guess_extension(cover_art.headers['content-type'])
-                    cover_art_output = os.path.join(self.cover_art_path, str(i) + extension)
-
-                    with open(cover_art_output, 'wb') as f:
-                        cover_art.raw.decode_content = True
-                        copyfileobj(cover_art.raw, f)
-
+                    cover_art_output = os.path.join(self.cover_art_path, "".join(x for x in album.title if x.isalnum()) + extension)
+                    if not os.path.exists(cover_art_output):
+                        with open(cover_art_output, 'wb') as f:
+                            cover_art.raw.decode_content = True
+                            copyfileobj(cover_art.raw, f)
                     cover_art_exists = True
                 else:
                     for key, path in COVER_ART_SUBS.items():
                         if album.title.lower().find(key) != -1:
                             extension = '.' + path.split('.')[-1]
-                            cover_art_output = os.path.join(self.cover_art_path, str(i) + extension)
+                            cover_art_output = os.path.join(self.cover_art_path, "".join(x for x in album.title if x.isalnum()) + extension)
 
                             with open(path, 'rb') as src:
                                 with open(cover_art_output, 'wb') as dest:
